backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
Pillow